Remove getVariableIndex
since the variables themselves will know their index.

This is OK because if there is a discrepancy between what the variables say their index is and their actual index, then one of the variables before it must have been deleted, which would occur only if its name was taken or its type is wrong, which means there is a problem in the source code and it shouldn't compile anyway.