Add LLVM binary instructions
Because AST binary expressions look totally different from each other in LLVM, I'm not grouping them under one 'LLVMBinary' family. They're just split into their three types: arithmetic, comparison, and logical.

The comparison LLVM instructions consist of AST equality and relational expressions. This is because the basic form is the same between them, the only difference is the type of the inputs. For equality, we would need to check what the types are. For relational, while we know the input types are LLVMIntType, we could generalize and make it easier for us to just look up the type in the value.