Add UID to nodes and LLVMRegisters
The UIDs are unique, increasing integer indices used when printing the nodes and registers out. They aren't set until after a topological sort of the nodes is created so that way the UID for the nodes will increase across the whole file, as opposed to only increasing in the order of creation. The same goes for the LLVMRegisters. That way node L0 will be the allocation node of the first function and L1 will be the node of the first non-allocation instructions of the function, and %u0 will not be the return register in the return instruction in the last node of the first function.