Use PrintWriter instead of System.out
When I realized last night that I wasn't going to use a filewriter, I neglected to think about using a temp file for Clang to read for compilation. Instead of using a BufferedWriter, however, I decided to go with the quick and dirty PrintWriter so I can use the println methods and I won't need to wrap everything in a try-catch.