Change LLVMParameter to LLVMLocal
In the LLVM code, in the allocation node, all parameters are pushed to the stack, meaning they get treated no differently than regular locals in the following nodes.

In addition to that, locals and parameters have different implicit types. Though we're not dealing with pointer arithmetic, in functions, parameters are their declared types (i32, i1, %struct.foo*), but locals are pointers to locations on the stack (i32*, i1*, %struct.foo**). Thus, only locals can be used in the context of a load instruction. The parameters exist only within the context of the allocation node. Once they're pushed to the stack, they aren't used ever again.