Add an encapsulated type for LLVMNull
In the case of calling a function with `null` as a parameter, that null doesn't know what it's type is. Now, when gathering the arguments for the function, if an argument is null, we will stick the type of the argument directly into the null value so the printer knows what type to print out.