Separate scanf scratch pointer and value
The issue that was occurring was that scanf requires an i32*, which is unique to only that function. Load also requires pointers (i32*, i1*, and %struct.foo**), but its pointers are implicit (that is, it appends a star to the end of the type to make it a pointer of the thing). This is because we aren't dealing with pointer arithmetic and all the structs behave nicely as just references to objects: we can just treat all the locals as their types until needing to load them, then we treat them like pointers.

It just sort of works out nicely, except for the scanf scratch type. Because scanf requires an i32*, the scanf scratch variable is an i32*; however, the load instruction needs an i32*, but to it it would be getting an i32**. Thus if we separated the variable into its pointer and value (which gets treated as a pointer), then it works out.

The alternative to this is to introduce new types for i32*, i1*, and %struct.foo** and ensure that all locals are of those types with methods to be able to convert from pointer to data type. Then we could have one variable for the scanf format, but we'd have twice as many types to manage.

The current system of implicitly treating locals as parameters works out because the locals appear only in loads and stores, which are the instructions that add stars to the types.