Change cleanCFG to not be recursive
Use data structures to hold onto which nodes to visit and which were visited. This way we won't get stack overflow on exceptionally deep CFGs.

Change removeUnreachables and removeEmpties to run separately instead of both in succession in the same loop. That way we don't run into the issue of removing empty nodes and accidentally keeping unreachable ones. I have a contrived and unrealistic scenario, but it can happen, possibly.

Make all those functions static.