Add generic armString to call lowerARMString
This function is like the llvmString in that every LLVM thing has one. All ARMValues are to have one. Because ints are constants and constants are values, immediate ints can be used as the second operand to some ARM instructions (like the mov family). As far as I can tell, the only immediate values that are used in instructions other than movw and movt are #0 and #1 for comparisons and boolean value checking. As such, we will use the lower 16 bits for those values.