Redefine LLVMValues
Before, we had distinguished between constants and variables, with parameters and (virtual) registers being lumped in together with locals and globals. This was fine for the most part because constant values could not change and variable values could. However, that thinking did not follow the fact that LLVM virtual registers must follow static single assignment and thus cannot change (even in an intrinsic way), unlike locals and globals, which can change (in memory).

I had the realization that I was thinking about the variable types wrong. The constants are all constant still, but parameters and virtual registers hold onto the values themselves, while globals and locals are pointers to memory or the stack (respectively). Thus I decided to split them up. Then I realized that constants and the register types can be used interchangably in many functions (add, for example, can take constants or registers). So the constants and registers are grouped together under LLVMOperand.