Move Getelementptr up a level
All the targeted instructions store their values into LLVMVirtual registers. Getelementptr gets a pointer to a location in memory, which must be dereferenced in Load or used as a destination in Store. As a result, it returns a value that is closer to an LLVMVariable than it is to an LLVMRegister. Thus, it moves up a level to outside the TargetedInstruction world and it gets its own target, much like Alloca.