Rewrite ARM values
In Assembly there are two data "types": registers and immediates. Of the immediates, sometimes they can be used as operands for instructions and sometimes they can't. Some immediates are small enough to fit in instructions and some are too large. Immediates that are too large to fit (generally larger than 8-bits, though some instructions allow for 12-bit immediates) must be put in registers, so they get classified as an ARMInteger. Globals and the STDIO "variables" are just pointers to memory and act just like 32-bit integers. Small immediate integers can sometimes be used as an operand in place of a register, so both registers and small constants are fit under ARMOperand.