Change Getelementptr to take a Virtual register
The source value for getelementptr will only ever be a virtual register. This instruction is used to dereference pointers to objects, which means they must be of struct type. There are no constant values of struct type (null is struct-equivalent or struct-comparable, but not struct type). So this type cast, though unchecked, will always be valid.