Add a scratch pointer
This will make doing function calls easier. The issue lies in the fact that, aside from the scratch pointer, all arguments are LLVMOperand types (constant values or registers, but not variable names). The scratch pointer is the only one that doesn't follow the pattern, so we create a separate one here, with similar functionality to the one in LLVMGlobal, but this acts as an LLVMPointerType.