Provide LLVMField with the base and offset
A field variable really is just pointer arithmetic on a base pointer that is held in a register, so make it just that. This will mean that we won't need to do any ARM equivalent of `getelementptr` (which would have been an addition), because we can use the register and offset value directly in ARM `ldr` and `str`.