Change UID rules for LLVMCFGNode
The first node of the graph will not have a UID assigned to it. It isn't needed since (in LLVM) the first node is immediately jumped into. The UID is assigned only if there is a jump into the first node, which is possible only if the first node is the guard of a loop.