Change `armString` to `armStrings`
Convert all armString functions to return a list of strings as opposed to just one string. This makes it easier to return nothing (in the cases where the instruction is equivalent to a no-op, such as a `mov r0, r0`. This also makes it easier to print extra lines in the case of register spillage.