Change order of stack frame management instructions
Previously, all stack locations were placed with respect to the stack pointer, so we would want to push clobbered registers before allocating space in the stack for locals. Now that all values are placed with respect to the frame pointer, we need to push clobbered registers after allocating space in the stack for locals.