Change default register number to 0
Now that we've got register allocation, no registers will print out that are virtual, so every register will get a number, with the default being 0.

There was previously an issue caused by how I designed my interference graph and register allocation function. The interference graph is build only by interference edges, so if a virtual register didn't interfere with anything, it would not be added to the graph. The register allocation function used the interference graph to allocate registers. Thus, if there was a register that didn't interfere with anything, it would not be allocated. The solution I've got now is to just allocate any unallocated registers to r0, which is fine because they won't conflict with anything else.