Loosen up Getelementptr to allow LLVMRegister
Now that we're doing SSA, we might be able to use a parameter value or a phi value as the source of a getelementptr instruction. I'm not sure if I'll loosen it further to include constants as well (so to go up to LLVMOperand). I don't think that would ever be needed.