Fail fast if a variable was not defined
All locals should be defined before use. Even parameters are defined. Globals are not sent into `readVariable` because they always go to memory.

If a call to `readVariable` results in `null`, then the function made it all the way up to the top node and couldn't find it along that path.