Add caching of ARMRegister in LLVMParameter
The `buildARM` method is called only once for each LLVMParameter in the stack-based compile, so we have no problem letting it cache a value it'll never use again. As for the register-based compile, it will be called constantly, so we will run the load once and cache the register used for it so subsequent accesses will access just the register and not perform a stack-load every time.